+++ src/hydrate/runner/index.ts
@@ -1,3 +1,4 @@
 export { createWindowFromHtml } from './create-window';
 export { hydrateDocument, renderToString, serializeDocumentToString, streamToString } from './render';
 export { setTagTransformer, transformTag } from '@platform';
+export { deserializeProperty, serializeProperty } from '@utils';